updated the Range Fan From Bearing and Traversal With Visibility tool
- Added layers for the Range fan and Visibility symbology
- Fix spelling errors
- Reinstated the Range Fan for final output
- Made the Output Coordinate system a user input, stating that is needs
to be a project system for proper processing.
- Changed the Delete field from “Visibility” for ”Gridcode” to keep the
visibility field in the final output
- Added “Traversal” Field in the Range Fan feature for additional
information
- Changed the “Minimum Bounding Geometry” from “Envelope” to
Convex_Hull” This change reduced the elevation foot print for
processing, reducing the multiple visibility returns received from the
previous version. Also changed the grouping from “All” to “None”. This
keep the elevation data for processing confined to the boundary of the
range fan.
- Change the “Buffer” process linear unit from 1KM to 30M, reducing
overlap of elevation processing information to reduce multiple
visibility returns, when there is obviously none. Changed the dissolve
type from “All” to “None” to keep the elevation data confined to the
specific range fan.
- Added project raster after extract by mask to ensure proper viewshed
processing. The input is the same user input at the start of the tool.