Minor adjustments
Made adjustments to some of the parameters with some of the tools.
Changed "UTDS" to "TDS" to reflect using the different levels of the TDS
datasets. Adjusted the Linear Obstacle parameters so the symbology is
applied at the end of the process. Made an adjustment to the "Raster
Off-Road Mobility" validation due to project packaging issues with the
tables links not remaining. Updated the Linear Obstacle lyrx file.