MAoT Additional Obstacle Rebuild
Rebuilt the MAoT Additional Obstacle tool. The tool did not take into
account whether if there was no feature type in the workspace. For
example, if there was no "Point" feature, the tool would fail to
continue. Now the tool works whether there is a type of feature or not
in the workspace, providing the necessary Additional Obstacle for
Deliberate HLZ Analysis tool.