Fixes to Count Incident by LOC tool
these changes fixed github issue #376. The changes to the 10.4 version
required changes within the model. In Pro this tool was actually working
correctly, other than after the tool ran, the .lyrx file was not being
referenced properly and the layer was not being added to the map with
the correct symbology... i went ahead and cleaned that up, so it should
now add to the map with the correct symbology.