Edits to Farthest on Circle (Pro Edition)
Farthest on Circle Tool in Maritime Decision Aid Tools.tbx was edited to
work in ArcGIS Pro. The original version (titled "Farthest On Circle -
Original" runs in about 10 seconds in ArcMap 10.4), runs in about 16
minutes in ArcGIS Pro. There are some issues with the extent variable
not being passed. Created another tool "Farthest on Circle" in the
"Maritime Decision Aid Toolbox" that works with ArcGIS Pro and uses a
different process. It goes much quicker, but still not as fast as ArcMap