Removed steps in the Table to Point model that are not need
The model has not been updated since ArcGIS for Desktop 10.1. At that
point the Convert Coordainte Notation GP tool did not keep the
attributes. Due to this there was a py script that was needed to add
unique ids to table and a join step to get your final output.

This is not needed anymore as of ArcGIS for Desktop 10.2