Changes to Incident Density tool
The Optimized Hot Spots tool that Incident Density uses does not like
using a geographic coordinate system. To make this tool more user
friendly the inputs are now projected before OHS runs, and the parameter
is added to the model to set the spatial reference once, for both
projects. This change was made for Pro and for arcmap