Updates to Clearing Operations and Compound Map
Action taken:
changed the script to rely on the script tool to set the symbology and
labels.
Note:
Based on best practices coming from the gp core team I have commented
out the steps that apply symbology in the script and set the script tool
to look at the layer file.