Added documentation from model tools onto script tools
In the course of stripping away the models and exposing the scripts as
tools, the documentation for the tools got wiped out. I have copied the
content from the old tools into the script tools. Note that these two
toolboxes are not named the same as the toolboxes in the ArcMap docs or
Pro project.  They match the solutions-geoprocessing-toolbox naming
convention.