Address issue #657
The script was reusing a rotate feature class function that was generic
so that it could be used against any feature class. As part of this it
used the arcpy.AddJoin_management to map over the existing feature class
attributes. This process seemed to cause an issue in Pro.

As we know that there is only one attrribute field to map over (Grid) I
have change script to do this in the search cursor and removed the join.